[MRESOLVER-153] Make TrackingFileManager shared singleton
Step one is to make TrackingFileManager a shared component,
also, using iface+defaultClass patter to add indirection possibilities
here as well.
