+++ maven-resolver-impl/src/main/java/org/eclipse/aether/internal/impl/EnhancedLocalRepositoryManagerFactory.java
@@ -19,6 +19,9 @@ package org.eclipse.aether.internal.impl;
  * under the License.
  */
 
+import java.util.Objects;
+
+import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
@@ -27,6 +30,8 @@ import org.eclipse.aether.repository.LocalRepository;
 import org.eclipse.aether.repository.LocalRepositoryManager;
 import org.eclipse.aether.repository.NoLocalRepositoryManagerException;
 import org.eclipse.aether.spi.localrepo.LocalRepositoryManagerFactory;
+import org.eclipse.aether.spi.locator.Service;
+import org.eclipse.aether.spi.locator.ServiceLocator;
 
 /**
  * Creates enhanced local repository managers for repository types {@code "default"} or {@code "" (automatic)}. Enhanced
@@ -38,13 +43,27 @@ import org.eclipse.aether.spi.localrepo.LocalRepositoryManagerFactory;
 @Singleton
 @Named( "enhanced" )
 public class EnhancedLocalRepositoryManagerFactory
-    implements LocalRepositoryManagerFactory
+    implements LocalRepositoryManagerFactory, Service
 {
     private float priority = 10.0f;
 
+    private TrackingFileManager trackingFileManager;
+
     public EnhancedLocalRepositoryManagerFactory()
     {
-        // enable no-arg constructor
+        // no arg ctor for ServiceLocator
+    }
+
+    @Inject
+    public EnhancedLocalRepositoryManagerFactory( final TrackingFileManager trackingFileManager )
+    {
+        this.trackingFileManager = Objects.requireNonNull( trackingFileManager );
+    }
+
+    @Override
+    public void initService( final ServiceLocator locator )
+    {
+        this.trackingFileManager = Objects.requireNonNull( locator.getService( TrackingFileManager.class ) );
     }
 
     public LocalRepositoryManager newInstance( RepositorySystemSession session, LocalRepository repository )
@@ -52,7 +71,7 @@ public class EnhancedLocalRepositoryManagerFactory
     {
         if ( "".equals( repository.getContentType() ) || "default".equals( repository.getContentType() ) )
         {
-            return new EnhancedLocalRepositoryManager( repository.getBasedir(), session );
+            return new EnhancedLocalRepositoryManager( repository.getBasedir(), session, trackingFileManager );
         }
         else
         {
