+++ maven-resolver-impl/src/test/java/org/eclipse/aether/internal/impl/DefaultArtifactResolverTest.java
@@ -272,7 +272,10 @@ public class DefaultArtifactResolverTest
         };
 
         repositoryConnectorProvider.setConnector( connector );
-        resolver.setUpdateCheckManager( new DefaultUpdateCheckManager().setUpdatePolicyAnalyzer( new DefaultUpdatePolicyAnalyzer() ) );
+        resolver.setUpdateCheckManager( new DefaultUpdateCheckManager()
+            .setUpdatePolicyAnalyzer( new DefaultUpdatePolicyAnalyzer() )
+            .setTrackingFileManager( new DefaultTrackingFileManager() )
+        );
 
         session.setResolutionErrorPolicy( new SimpleResolutionErrorPolicy( true, false ) );
         session.setUpdatePolicy( RepositoryPolicy.UPDATE_POLICY_NEVER );
